Extract SubselectRewriter out of RelationNormalizer
This will allow us to test these components in isolation.
In addition the Context class became unnecessary in the
RelationNormalizer - The EvaluatingNormalizer is now created once per
startup instead of per normalize.
